requirements.txt
PyOpenGL
PyOpenGL-accelerate
Pillow
pygame